Splits the ImagePublisher into CameraImageManager(attached to camera object) and CameraImagePublisher (attached to the ROSConnector)
