+++ Unity3D/Assets/RosSharp/Scripts/CameraImageManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 3fbd0120f43654542ac5e331a54463a7
-timeCreated: 1512548392
+guid: 079cb956cf1dc40e9a60481f68e8639d
+timeCreated: 1512650540
 licenseType: Free
 MonoImporter:
   externalObjects: {}
