+++ Unity3D/Assets/RosSharp/Scripts/CameraImagePublisher.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: b6e421c2399834365b860e6e892efd9f
+timeCreated: 1512650730
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
